Add utility functions to the base test class
The functions allow to return a String representation of the properties
in the test instance, or to load properties into the instace from a
String. These functions leverage the Properties.store and .load
respectively, thus allowing in a single statement to call each method
while avoiding the String to InputStream or OutputStream conversion and
I/O exception handling boilerplate.