Fix some Properties.load tests
String values under test are updated to match the description comment,
in test methods for property lines ending with a backslash.

Names of test methods about backslash-escaped characters in property
values are updated to be clearer (although longer).

A test description comment is updated to convey the non-comment nature
of bang-led lines.