Rework slightly Properties.getProperty test
The property key is extracted in a local variable, to signify more
clearly that it is the same used throughout the test.
Also, the assertion check is now of equality, not identity: the method
does not guarantee that it will return the same value that was set, only
that it is equal.