Cover both load/store overloads in tests
The utility methods to load properties from a String and store
properties into a String are removed from the base test class, and
redeclared as abstract methods in the relevant test case. Two subclasses
are added for each case, each implementing the
loadFromString/storeToString method using either a stream or
reader/writer object. Since test methods are inherited, this allows to
run the tests twice, with both overloads, without duplicating the test
code.