Fix line-unwrapping internal method
The method now returns the unwrapped line, instead of reassigning its
argument (overlooked proofreading the auto-refactoring).
The call instruction to the method is also modified, to make use of the
returned string: it is assigned to the variable holding the current
line.
Also, a redundant control flow statement is removed (the continue
keyword before the end of a loop).