Implement unescape method
The method converts the given escaped character to the one for which it
stands, simply through a big switch block.
If the character does not fall in one of the cases, an exception is
raised.
Also, the boolean parameter indicating whether the method is called
from within the key or the value is removed, since the two contexts
allow the same sequences.