Fix test ensuring that octal form is not handled
The first digit of the octal sequence was a zero. This construct raised
no error when parsed, as backslash-zero is an authorized escape
sequence, however it does not stand for an octal number. To ensure that
the test does what is intended, this digit is changed to one that does
not constitute a valid escape sequence.