Refactor Properties class
The method skipLeadingWhitespace is modified to accept a starting index,
and thus renamed to match (since it no longer skips leading whitespace,
but any whitespace within the given string). The occurrences of the
original method are replaced with calls to the new one with a starting
index of 0.

The method extractKey, that was also used in extractVazlue, is given a
more generic name (not bound to the concept of the property key). The
latter method is also inlined into its only call site, since by reusing
the whitespace-skipping method it is reduced to a single statement.