Introduce line number field
The field will be used to display in error messages the number of the
line at which the error was detected.

Currently, the field is not used. It is set to zero at the beginning of
the overload of method load containing the actual implementation, and
incremented each time a line is read from the input.