Slightly improve method extractComponent
The string builder variable is renamed to match the property
component-agnostic identifier of the method. While loop using
Character.isWhitespace method to test a range of whitespace characters
is replaced with a call to the dedicated utility method.
Also, an indentation is fixed in the internal overload of method load