Also process value when writing
The property writing of the property value is also extracted into its
own function. It is notably more complex, as there are more characters
who need escaping, and because the processing of the first and last
characters is different (these need to be escaped too if they are a
space, to not be discarded as filler whitespace). Only different in a
detail, but enough to make the merge of the corresponding code blocks
impossible (can't refactor only a case in a switch, at least not without
introducing inopportune complications e.g. in the form of a boolean
parameter).