Add more tests for Properties.load
One test ensures that the method does not accept control characters in
property keys. In practice, it checks for the NUL byte (ideally, all
control char escape sequences would be checked).
The other two tests ensure that the method respectively forbid (by means
of exception throwing) or accepts an empty key or value.