Enforce stricter requirements on the key
Escaped leading/trailing whitespace is no longer accepted, nor are
control characters.
This is implemented by adding a parameter "isKey" to the method
extractComponent, that, when true, enables additional checks on
unescaped characters. This is certainly not the best way of doing it,
and may be an antipattern... It will have to do for now.